+++ MOP/src/RuleFiles/IgnoreRule.cs
@@ -57,5 +57,6 @@ namespace MOP
     {
         public bool SatsumaIgnoreRenderers;
         public bool DontDestroyEmptyBeerBottles;
+        public bool ExperimentalSatsumaFix;
     }
 }